Add fetch functions for subscription resource
