+++ front/pages/api/login.ts
@@ -25,8 +25,8 @@ import {
 } from "@app/lib/iam/workspaces";
 import type { MembershipInvitation } from "@app/lib/models/membership_invitation";
 import { Workspace } from "@app/lib/models/workspace";
-import { subscriptionForWorkspace } from "@app/lib/plans/subscription";
 import { MembershipResource } from "@app/lib/resources/membership_resource";
+import { SubscriptionResource } from "@app/lib/resources/subscription_resource";
 import type { UserResource } from "@app/lib/resources/user_resource";
 import { getSignUpUrl } from "@app/lib/signup";
 import { ServerSideTracking } from "@app/lib/tracking/server";
@@ -270,7 +270,7 @@ async function handleRegularSignupFlow(
       );
     }
 
-    const workspaceSubscription = await subscriptionForWorkspace(
+    const workspaceSubscription = await SubscriptionResource.fetchByWorkspace(
       renderLightWorkspaceType({ workspace: existingWorkspace })
     );
     const hasAvailableSeats = await evaluateWorkspaceSeatAvailability(
